requirements.txt
# === MIDAS Voice Assistant Dependencies ===
# See README.md for full installation instructions

# Core Web Framework
fastapi
uvicorn[standard]
python-multipart
websockets

# Speech-to-Text (Whisper)
faster-whisper

# Large Language Model
# NOTE: Install with CUDA support manually:
#   Windows: $env:CMAKE_ARGS="-DGGML_CUDA=on"; pip install llama-cpp-python --force-reinstall
#   Linux:   CMAKE_ARGS="-DGGML_CUDA=on" pip install llama-cpp-python --force-reinstall
llama-cpp-python

# Text-to-Speech (Silero - model downloaded via download_models.py)
# PyTorch with CUDA - Install manually FIRST:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch

# Audio Processing
soundfile
numpy

# Model Downloading (optional, download_models.py uses urllib instead)
# huggingface-hub
# tqdm